Moving classes to separate files.
