+++ test/test_encode.py
@@ -1,0 +1,7 @@
+import unittest
+from monerorequest.encode import Encode
+
+class TestEncode(unittest.TestCase):
+    def test_invalid_payment_request_encoding(self):
+        with self.assertRaisesRegex(ValueError, 'Invalid input'):
+            Encode.monero_payment_request_from_json(json_data='', version='')
